+++ MindEcho/MindEcho/Views/TranscriptionView.swift
@@ -1,7 +1,6 @@
 import MindEchoCore
 import Speech
 import SwiftUI
-import UIKit
 
 struct TranscriptionView: View {
     let recording: Recording
@@ -91,70 +90,16 @@ struct TranscriptionView: View {
     private func sharePDF() {
         guard case .success(let text) = viewModel.state else { return }
         let title = "書き起こし #\(recording.sequenceNumber)"
-        let pageRect = CGRect(x: 0, y: 0, width: 595.2, height: 841.8)
-        let margin: CGFloat = 50
-        let contentRect = pageRect.insetBy(dx: margin, dy: margin)
-
-        let titleAttributes: [NSAttributedString.Key: Any] = [
-            .font: UIFont.boldSystemFont(ofSize: 20),
-            .foregroundColor: UIColor.label,
-        ]
-        let bodyAttributes: [NSAttributedString.Key: Any] = [
-            .font: UIFont.systemFont(ofSize: 14),
-            .foregroundColor: UIColor.label,
-        ]
-
-        let renderer = UIGraphicsPDFRenderer(bounds: pageRect)
-        let data = renderer.pdfData { context in
-            context.beginPage()
-            let titleString = NSAttributedString(string: title, attributes: titleAttributes)
-            let titleSize = titleString.boundingRect(
-                with: CGSize(width: contentRect.width, height: .greatestFiniteMagnitude),
-                options: [.usesLineFragmentOrigin],
-                context: nil
-            )
-            titleString.draw(in: CGRect(
-                x: contentRect.origin.x,
-                y: contentRect.origin.y,
-                width: contentRect.width,
-                height: titleSize.height
-            ))
-
-            let bodyTop = contentRect.origin.y + titleSize.height + 20
-            let bodyString = NSAttributedString(string: text, attributes: bodyAttributes)
-            let textStorage = NSTextStorage(attributedString: bodyString)
-            let layoutManager = NSLayoutManager()
-            textStorage.addLayoutManager(layoutManager)
-
-            var currentY = bodyTop
-            var glyphIndex = 0
-
-            while glyphIndex < layoutManager.numberOfGlyphs {
-                let availableHeight = pageRect.height - margin - currentY
-                let textContainer = NSTextContainer(
-                    size: CGSize(width: contentRect.width, height: availableHeight)
-                )
-                textContainer.lineFragmentPadding = 0
-                layoutManager.addTextContainer(textContainer)
-
-                let glyphRange = layoutManager.glyphRange(for: textContainer)
-                if glyphRange.length == 0 { break }
-
-                layoutManager.drawBackground(forGlyphRange: glyphRange, at: CGPoint(x: contentRect.origin.x, y: currentY))
-                layoutManager.drawGlyphs(forGlyphRange: glyphRange, at: CGPoint(x: contentRect.origin.x, y: currentY))
-
-                glyphIndex = NSMaxRange(glyphRange)
-                if glyphIndex < layoutManager.numberOfGlyphs {
-                    context.beginPage()
-                    currentY = margin
-                }
-            }
-        }
+        let data = PDFRenderer.render(title: title, body: text)
 
         let tempURL = FileManager.default.temporaryDirectory
             .appendingPathComponent("transcription_\(recording.sequenceNumber).pdf")
-        try? data.write(to: tempURL)
-        shareItems = [tempURL]
+        do {
+            try data.write(to: tempURL)
+            shareItems = [tempURL]
+        } catch {
+            // PDF write failed — do not present share sheet with invalid URL
+        }
     }
 
     @ViewBuilder
